refactor: move VoiceSynth to shared module
