+++ shared/ts/src/index.ts
@@ -18,4 +18,5 @@ export * from './migrations/integrity.js';
 export * from './migrations/contract.js';
 export * from './providers/discord/normalize.js';
 export * from './web/url.js';
-export * from "./stream/title.js";
+export * from './stream/title.js';
+export * from './voice-synth.js';
